requirements-minimal.txt
fastapi==0.70.0
uvicorn==0.15.0
neo4j==5.27.0
pydantic==1.10.19
numpy>=1.23.0,<2.0.0
pandas>=1.5.2,<2.0.0
requests==2.28.1
certifi==2023.7.22
python-dotenv==1.0.0
